+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # tarka
 
+## 0.8.0
+- key argument added to iterator merge utilities
+- iterator merge utilities can handle descending order as well
+
 ## 0.7.0
 - NamedObject utility (elegant singleton objects)
 - merge-zip utility with iterators